extract getAnnotions to be used in both components
